rename `dev.py` to `main.py`
